scale to height. And scale to terminal size (unix only).
The exact semantics of the -width flag is somewhat different, although
backwards compatible.  It specifies a _maximum_ width as opposed to an
exact width for the output.  Similarly, the -height flag specifies a
maximum height.

The -scale flag overrides -width and -height and uses the invoking
terminal's width and height as guides.