render full frames for GIF images (broken)
Ensures the frame size is always the logical screen size.  The
background index is made transparent in the global color map and the
transparent index (which may change from frame to frame) is only
logically transparent and not physically changed to be transparent in
the palette.

Rendering of frames attempts to obey the disposal flags in the Graphic
Control Extension block (GIF89a section 23).

- When unspecified an uninitialized transparent frame background is
  used.
- When "do not dispose" the last "do not dispose" frame is recalled and
  the rendered frame is saved.
- When "restore background" a transparent frame background is used.
- When "restore previous" the previous "do not dispose" frame is
  recalled.  a transparent background is used when no such frame exists.

Tests are failing.